Redesign thread synchronization logic
The same mutex was being used for two distinct situation. First, the API
functions must not run at the same time as the bus watch and playback
timer. Second, the stream EOS probe was modifying the current and next
streams.

Get rid of the second situation by _not_ updating the streams in the probe,
and instead setting a flag that an update is necessary. As for the first,
rename m_mutex to m_loop_mutex to make its purpose clearer.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>