Rework buffering to allow for custom low/high thresholds and prioritization
The current stream now has priority. If it needs to buffer, and the next
stream is also buffering, then the next stream's buffering is blocked
until the current stream is done.

The buffering parameters are redefined as size, timeout, estimation
duration, allowing for better control.

Custom low/high percentage is now also possible, defining when the
buffering state starts and ends.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>