Don't switch to paused state after starting; switch to playing directly
The switch to paused is unnecessary. It would just immediately switch
further to playing anyway, causing a sequence starting->paused->playing.
Just switch to the PLAYING GStreamer state directly. The PLAYING handler
in the bus watch will then switch the pipeline state to playing as well.
Result is a direct sequence starting->playing.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>