Use the correct mutex in the sync handler
m_current_stream is changed in the set_pipeline_to_idle() function
after the pipeline state has been switched to NULL, in play_media()
after the pipeline got reinitialized, and in the
make_next_stream_current_nolock() function, which is protected by the
m_stream_mutex. Therefore, the m_loop_mutex at line 1870 is not needed,
and can be replaced by the m_stream_mutex. This fixes a potential deadlock
when set_gstreamer_state_nolock() is called while the sync handler is
running. (This can be provoked by running the cmdline player with an
invalid URL as argument.)

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>