Add configuration ability for RTSP and packet loss callback
It is now possible to instruct the pipeline to set up RTSP connections in
a specific way. For example, there is now the option to use TCP instead of
UDP as the media transport for RTSP. It is also now possible to configure
packet retransmissions and the length of the RTP jitter buffer.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>